requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
colormath==3.0.0
contourpy==1.3.1
cycler==0.12.1
dlib==19.24.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.0
greenlet==3.1.1
imageio==2.36.0
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
networkx==3.4.2
numpy==2.1.3
opencv-contrib-python==4.10.0.84
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
tifffile==2024.9.20
typing_extensions==4.12.2
Werkzeug==3.1.3
